03-08-2020 : 20:27

+++ frontEnd/src/app/components/folders/folders.component.ts
@@ -3,6 +3,8 @@ import { FolderModel } from "../../models/folder-model";
 import { FolderServiceService } from "../../services/folder-service.service";
 import { ActivatedRoute, Router, ParamMap} from "@angular/router";
 
+declare var Swal: any;
+
 @Component({
   selector: 'app-folders',
   templateUrl: './folders.component.html',
@@ -25,8 +27,27 @@ export class FoldersComponent implements OnInit {
   }
 
   delete(folder){
-    this.folderService.deleteFolder(folder).subscribe(
-      res => { this.ngOnInit(); },
-      error => { console.log(error); });
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.folderService.deleteFolder(folder).subscribe(
+          res => {
+            this.ngOnInit();
+            Swal.fire(
+              'Deleted!',
+              'Your folder has been deleted.',
+              'success'
+            );
+          },
+          error => { console.log(error); });
+      }
+    });
   }
 }
